+++ build.gradle
@@ -23,8 +23,6 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springframework.boot:spring-boot-starter-data-rest'
-    implementation 'org.springframework.boot:spring-boot-starter-webflux'
-    implementation 'io.netty:netty-resolver-dns-native-macos:4.1.111.Final'
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.0.4'
 
     implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
